requirements.txt
openai
fastapi
uvicorn
python-multipart
numpy
opencv-python-headless
ultralytics
comet_ml
pandas
streamlit
roboflow
pyyaml
